+++ src/routers/Router.js
@@ -6,7 +6,7 @@ import ShopGuideDetails from '../pages/ShopGuideDetails.jsx';
 import SignIn from '../pages/SignIn.jsx';
 import SignUp from '../pages/SignUp.jsx';
 import ShopGuidePosting from '../pages/ShopGuidePosting.jsx';
-
+import ShopGuideArticle from '../pages/ShopGuideArticle.jsx';
 
 const Router = () => {
   return (
@@ -16,7 +16,12 @@ const Router = () => {
         <Route exact path='/' element={<Home />} />
         <Route path='/signin' element={<SignIn />} />
         <Route path='/shopguide' element={<ShopGuide />} />
-        <Route path='/shopguidedetails/:id' component={ShopGuideDetails} element={<ShopGuideDetails />} />
+        <Route path='/shopguidearticle' element={<ShopGuideArticle />} />
+        <Route
+          path='/shopguidedetails/:id'
+          component={ShopGuideDetails}
+          element={<ShopGuideDetails />}
+        />
         <Route path='/signup' element={<SignUp />} />
         <Route path='/shopguideposting' element={<ShopGuidePosting />} />
       </Routes>
